Mule 13492 (#5695)
MULE-13492: Fixes for building with Java 8 target configuration

+++ tests/integration/src/test/java/org/mule/SessionPropertiesInExceptionStrategyTestCase.java
@@ -42,6 +42,6 @@ public class SessionPropertiesInExceptionStrategyTestCase extends FunctionalTest
 
         assertNull(result.getExceptionPayload());
         assertFalse(result.getPayload() instanceof NullPayload);
-        assertEquals(list.size(), result.getSessionProperty("ErrorCount"));
+        assertEquals(list.size(), (int)result.getSessionProperty("ErrorCount"));
     }
 }
